+++ wgpu-core/Cargo.toml
@@ -167,6 +167,7 @@ wgpu-core-deps-windows-linux-android = { workspace = true, optional = true }
 
 [dependencies]
 naga.workspace = true
+wgpu-naga-bridge.workspace = true
 wgpu-hal.workspace = true
 wgpu-types.workspace = true
