+++ wgpu-core/src/device/mod.rs
@@ -379,203 +379,5 @@ impl WebGpuError for MissingDownlevelFlags {
     }
 }
 
-/// Compute naga [`Capabilities`] corresponding to [`Features`] and [`DownlevelFlags`].
-///
-/// [`Capabilities`]: naga::valid::Capabilities
-/// [`Features`]: wgt::Features
-/// [`DownlevelFlags`]: wgt::DownlevelFlags
-pub fn features_to_naga_capabilities(
-    features: wgt::Features,
-    downlevel: wgt::DownlevelFlags,
-) -> naga::valid::Capabilities {
-    use naga::valid::Capabilities as Caps;
-    let mut caps = Caps::empty();
-    caps.set(
-        Caps::IMMEDIATES,
-        features.contains(wgt::Features::IMMEDIATES),
-    );
-    caps.set(Caps::FLOAT64, features.contains(wgt::Features::SHADER_F64));
-    caps.set(
-        Caps::SHADER_FLOAT16,
-        features.contains(wgt::Features::SHADER_F16),
-    );
-    caps.set(
-        Caps::SHADER_FLOAT16_IN_FLOAT32,
-        downlevel.contains(wgt::DownlevelFlags::SHADER_F16_IN_F32),
-    );
-    caps.set(
-        Caps::PRIMITIVE_INDEX,
-        features.contains(wgt::Features::SHADER_PRIMITIVE_INDEX),
-    );
-    caps.set(
-        Caps::TEXTURE_AND_SAMPLER_BINDING_ARRAY,
-        features.contains(wgt::Features::TEXTURE_BINDING_ARRAY),
-    );
-    caps.set(
-        Caps::BUFFER_BINDING_ARRAY,
-        features.contains(wgt::Features::BUFFER_BINDING_ARRAY),
-    );
-    caps.set(
-        Caps::STORAGE_TEXTURE_BINDING_ARRAY,
-        features.contains(wgt::Features::TEXTURE_BINDING_ARRAY)
-            && features.contains(wgt::Features::STORAGE_RESOURCE_BINDING_ARRAY),
-    );
-    caps.set(
-        Caps::STORAGE_BUFFER_BINDING_ARRAY,
-        features.contains(wgt::Features::BUFFER_BINDING_ARRAY)
-            && features.contains(wgt::Features::STORAGE_RESOURCE_BINDING_ARRAY),
-    );
-    caps.set(
-        Caps::TEXTURE_AND_SAMPLER_BINDING_ARRAY_NON_UNIFORM_INDEXING,
-        features
-            .contains(wgt::Features::SAMPLED_TEXTURE_AND_STORAGE_BUFFER_ARRAY_NON_UNIFORM_INDEXING),
-    );
-    caps.set(
-        Caps::BUFFER_BINDING_ARRAY_NON_UNIFORM_INDEXING,
-        features.contains(wgt::Features::UNIFORM_BUFFER_BINDING_ARRAYS),
-    );
-    caps.set(
-        Caps::STORAGE_TEXTURE_BINDING_ARRAY_NON_UNIFORM_INDEXING,
-        features.contains(wgt::Features::STORAGE_TEXTURE_ARRAY_NON_UNIFORM_INDEXING),
-    );
-    caps.set(
-        Caps::STORAGE_BUFFER_BINDING_ARRAY_NON_UNIFORM_INDEXING,
-        features
-            .contains(wgt::Features::SAMPLED_TEXTURE_AND_STORAGE_BUFFER_ARRAY_NON_UNIFORM_INDEXING),
-    );
-    caps.set(
-        Caps::ACCELERATION_STRUCTURE_BINDING_ARRAY,
-        features.contains(wgt::Features::ACCELERATION_STRUCTURE_BINDING_ARRAY),
-    );
-    caps.set(
-        Caps::STORAGE_TEXTURE_16BIT_NORM_FORMATS,
-        features.contains(wgt::Features::TEXTURE_FORMAT_16BIT_NORM),
-    );
-    caps.set(Caps::MULTIVIEW, features.contains(wgt::Features::MULTIVIEW));
-    caps.set(
-        Caps::EARLY_DEPTH_TEST,
-        features.contains(wgt::Features::SHADER_EARLY_DEPTH_TEST),
-    );
-    caps.set(
-        Caps::SHADER_INT64,
-        features.contains(wgt::Features::SHADER_INT64),
-    );
-    caps.set(
-        Caps::SHADER_INT64_ATOMIC_MIN_MAX,
-        features.intersects(
-            wgt::Features::SHADER_INT64_ATOMIC_MIN_MAX | wgt::Features::SHADER_INT64_ATOMIC_ALL_OPS,
-        ),
-    );
-    caps.set(
-        Caps::SHADER_INT64_ATOMIC_ALL_OPS,
-        features.contains(wgt::Features::SHADER_INT64_ATOMIC_ALL_OPS),
-    );
-    caps.set(
-        Caps::TEXTURE_ATOMIC,
-        features.contains(wgt::Features::TEXTURE_ATOMIC),
-    );
-    caps.set(
-        Caps::TEXTURE_INT64_ATOMIC,
-        features.contains(wgt::Features::TEXTURE_INT64_ATOMIC),
-    );
-    caps.set(
-        Caps::SHADER_FLOAT32_ATOMIC,
-        features.contains(wgt::Features::SHADER_FLOAT32_ATOMIC),
-    );
-    caps.set(
-        Caps::MULTISAMPLED_SHADING,
-        downlevel.contains(wgt::DownlevelFlags::MULTISAMPLED_SHADING),
-    );
-    caps.set(
-        Caps::DUAL_SOURCE_BLENDING,
-        features.contains(wgt::Features::DUAL_SOURCE_BLENDING),
-    );
-    caps.set(
-        Caps::CLIP_DISTANCE,
-        features.contains(wgt::Features::CLIP_DISTANCES),
-    );
-    caps.set(
-        Caps::CUBE_ARRAY_TEXTURES,
-        downlevel.contains(wgt::DownlevelFlags::CUBE_ARRAY_TEXTURES),
-    );
-    caps.set(
-        Caps::SUBGROUP,
-        features.intersects(wgt::Features::SUBGROUP | wgt::Features::SUBGROUP_VERTEX),
-    );
-    caps.set(
-        Caps::SUBGROUP_BARRIER,
-        features.intersects(wgt::Features::SUBGROUP_BARRIER),
-    );
-    caps.set(
-        Caps::RAY_QUERY,
-        features.intersects(wgt::Features::EXPERIMENTAL_RAY_QUERY),
-    );
-    caps.set(
-        Caps::SUBGROUP_VERTEX_STAGE,
-        features.contains(wgt::Features::SUBGROUP_VERTEX),
-    );
-    caps.set(
-        Caps::RAY_HIT_VERTEX_POSITION,
-        features.intersects(wgt::Features::EXPERIMENTAL_RAY_HIT_VERTEX_RETURN),
-    );
-    caps.set(
-        Caps::TEXTURE_EXTERNAL,
-        features.intersects(wgt::Features::EXTERNAL_TEXTURE),
-    );
-    caps.set(
-        Caps::SHADER_BARYCENTRICS,
-        features.intersects(wgt::Features::SHADER_BARYCENTRICS),
-    );
-    caps.set(
-        Caps::MESH_SHADER,
-        features.intersects(wgt::Features::EXPERIMENTAL_MESH_SHADER),
-    );
-    caps.set(
-        Caps::MESH_SHADER_POINT_TOPOLOGY,
-        features.intersects(wgt::Features::EXPERIMENTAL_MESH_SHADER_POINTS),
-    );
-    caps.set(
-        Caps::COOPERATIVE_MATRIX,
-        features.intersects(wgt::Features::EXPERIMENTAL_COOPERATIVE_MATRIX),
-    );
-    caps.set(
-        Caps::PER_VERTEX,
-        features.intersects(wgt::Features::SHADER_PER_VERTEX),
-    );
-    caps.set(
-        Caps::DRAW_INDEX,
-        features.intersects(wgt::Features::SHADER_DRAW_INDEX),
-    );
-    caps.set(
-        Caps::MEMORY_DECORATION_COHERENT,
-        features.contains(wgt::Features::MEMORY_DECORATION_COHERENT),
-    );
-    caps.set(
-        Caps::MEMORY_DECORATION_VOLATILE,
-        features.contains(wgt::Features::MEMORY_DECORATION_VOLATILE),
-    );
-
-    caps
-}
-
-/// Create a validator for Naga [`Module`]s.
-///
-/// Create a Naga [`Validator`] that ensures that each [`naga::Module`]
-/// presented to it is valid, and uses no features not included in
-/// `features` and `downlevel`.
-///
-/// The validator can only catch invalid modules and feature misuse
-/// reliably when the `flags` argument includes all the flags in
-/// [`ValidationFlags::default()`].
-///
-/// [`Validator`]: naga::valid::Validator
-/// [`Module`]: naga::Module
-/// [`ValidationFlags::default()`]: naga::valid::ValidationFlags::default
-pub fn create_validator(
-    features: wgt::Features,
-    downlevel: wgt::DownlevelFlags,
-    flags: naga::valid::ValidationFlags,
-) -> naga::valid::Validator {
-    let caps = features_to_naga_capabilities(features, downlevel);
-    naga::valid::Validator::new(flags, caps)
-}
+pub use wgpu_naga_bridge::create_validator;
+pub use wgpu_naga_bridge::features_to_naga_capabilities;
